@@ -1,4 +1,4 @@
-package com.knoyo.slidedone.example
+package com.example.slide_done.example_slide_done
 
 import io.flutter.embedding.android.FlutterActivity
 
